sphinx_rtd_dark_mode_v2/pyproject.toml
[project]
name = "sphinx-rtd-dark-mode"
version = "1.2.4"
description = "Dark mode for the Sphinx Read the Docs theme."
authors = [
    {name = "MrDogeBro"},
    {name = "Uday Vidyadharan"}

]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.6"


classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# Requirements: This is done differently by poetry!
dependencies = [
   "sphinx-rtd-theme", 
   "Pillow>=9.2.0"
]

[project.urls]
repository = "https://github.com/FIRST-Tech-Challenge/ftcdocs-helper/tree/main/sphinx_rtd_dark_mode_v2"

[build-system]
requires = [
    "setuptools >= 35.0.2",
    "setuptools_scm >= 2.0.0, <3",
    "sphinx-rtd-theme",
    "Pillow>=9.2.0"
]

[tool.setuptools]
packages = ["sphinx_rtd_dark_mode"]